Add files via upload
Here I put up three examples to demonstrate both the forward and backward propagation process of simple neural network.The first one is using input with a rank 1 object or when you use zip(x, y) to iterate through the samples. The problem with rank 1 array is when you do the calculation, the dimension of the operation become subtle to track which may end up with bugs you do not know. This is the reason why Andrew NG specifically addressed that you should not use rank 1 array during modeling. 

Hence the second example is an improvement by changing any rank 1 array to rank 2 array.

The third example goes with vectorized sample training which is what really happens in real world.